fix(pdf): honor lexical xref stream boundaries
Consume the mandatory per-row algorithm tag for every PNG predictor, require a whitespace boundary after recovered xref keywords, reject declared xref fields wider than Long, and locate stream payloads through the lexer so comments and long whitespace cannot redirect parsing.

Use Long arithmetic for tile edges before clamping to prevent saturated Int dimensions from wrapping negative. Validation: the focused suites passed all 28 tests after failing exactly the five second-round regressions; the full suite also passed.

+++ src/main/kotlin/dev/jaeyoung/fileloom/pdf/render/PdfRenderPolicy.kt
@@ -76,8 +76,12 @@ object PdfRenderPolicy {
         }
         val left = tile.x.coerceIn(0, renderWidth)
         val top = tile.y.coerceIn(0, renderHeight)
-        val right = (tile.x + tile.width).coerceIn(left, renderWidth)
-        val bottom = (tile.y + tile.height).coerceIn(top, renderHeight)
+        val right = (tile.x.toLong() + tile.width.toLong())
+            .coerceIn(left.toLong(), renderWidth.toLong())
+            .toInt()
+        val bottom = (tile.y.toLong() + tile.height.toLong())
+            .coerceIn(top.toLong(), renderHeight.toLong())
+            .toInt()
         val width = right - left
         val height = bottom - top
         if (width <= 0 || height <= 0) return null
